PR-E-9: extract ClosureEmitter (closure orchestration + display-class synthesis)
Moves the closure-environment orchestration and display-class synthesis
surface out of ReflectionMetadataEmitter into a new internal sealed
ClosureEmitter at src/Core/CodeAnalysis/Emit/ClosureEmitter.cs. This
PR is the structural home for the eventual #567 + #503 unifying fix
(closure-environment emission unified with display-class emission); the
fix itself is deferred to PR-E-11 after BodyEmitter is promoted to
MethodBodyEmitter.

Approach: Option B — same playbook PR-E-5 ConversionEmitter used.
Top-level orchestration moves into ClosureEmitter; the BodyEmitter-
internal closure-emit methods stay on BodyEmitter and move with it
in PR-E-11. Option A would require widening BodyEmitter's private
surface to expose this.il / this.outer / this.locals / this.parameters
through an IBodyEmitContext interface only to take it apart again in
PR-E-11.

Moved into ClosureEmitter
-------------------------
Nested classes (was ReflectionMetadataEmitter.cs ~7125-7280):
  * ClosureInfo
  * CaptureRewriter : BoundTreeRewriter
  * ConstructedTypeCollector : BoundTreeRewriter  (deferred to here
    from PR-E-4 SlotPlanner per its explicit deferral note)

Top-level orchestration methods (was ~4789, ~4812, ~5588):
  * SynthesizeClosures
  * SynthesizeGoClosures
  * SynthesizeDisplayClass
  * IsTaskClrType (private helper; widened to internal static so the
    still-on-RME BodyEmitter.EmitGoStatement path can call it; both
    call sites move in PR-E-11)

Closure caches (was ~106, ~110, ~111, ~120, ~123):
  * ClosureInfos                  (was closureInfos)
  * GoClosureInfos                (was goClosureInfos)
  * SynthesizedClosureClasses     (was synthesizedClosureClasses)
  * ClosureInvokeToInfo           (was closureInvokeToInfo)
  * Counter                       (was closureCounter; public field
    so the still-on-RME state-machine synthesizers can pass it to
    Interlocked.Increment by ref, matching the pre-refactor shape)

Stayed on ReflectionMetadataEmitter (per plan)
----------------------------------------------
  * lambdaBodies dictionary — cross-cutting, populated by closures
    AND by state-machine synthesizers. RME owns it; ClosureEmitter
    receives the dictionary reference via constructor injection and
    writes through it from SynthesizeDisplayClass. The plan nominated
    this option ("Pick (a) for this PR").
  * RegisterConstructedTypeAliases — RME-side caller of the moved
    ConstructedTypeCollector. Updated to call
    new ClosureEmitter.ConstructedTypeCollector().
  * BodyEmitter-internal closure-emit methods — EmitFunctionLiteral
    (both overloads), EmitMethodGroup, EmitFunctionToDelegateConversion,
    EmitFunctionToNamedDelegateConversion,
    EmitFunctionLiteralToNamedDelegate, EmitMethodGroupToNamedDelegate,
    EmitClrEventSubscription, EmitUserEventSubscription,
    EmitCapturedVariableLoad, EmitOpenDelegateDynamicInvoke. Deferred
    to PR-E-11 MethodBodyEmitter (Option B).
  * State-machine synthesizers (SynthesizeIteratorStateMachines,
    SynthesizeAsyncIteratorStateMachines,
    SynthesizeAsyncLambdaStateMachines) and IteratorStateMachineInfo /
    AsyncIteratorEmitContext nested classes — move with PR-E-10
    StateMachineEmitter.

Wiring
------
ReflectionMetadataEmitter.EmitCore now instantiates ClosureEmitter
after TypeDefEmitter, passing emitCtx, cache, wellKnown, slotPlanner
and the shared lambdaBodies dictionary. Internal call sites updated:
  * this.SynthesizeClosures(...) → this.closures.SynthesizeClosures(...)
  * this.SynthesizeGoClosures(...) → this.closures.SynthesizeGoClosures(...)
  * this.closureInfos[...] → this.closures.ClosureInfos[...]
  * this.goClosureInfos[...] → this.closures.GoClosureInfos[...]
  * this.synthesizedClosureClasses → this.closures.SynthesizedClosureClasses
  * this.closureInvokeToInfo[...] → this.closures.ClosureInvokeToInfo[...]
  * ref this.closureCounter → ref this.closures.Counter
  * this.outer.closureInfos / .goClosureInfos (from BodyEmitter) →
    this.outer.closures.ClosureInfos / .GoClosureInfos
  * BodyEmitter.enclosingClosure field/ctor-param type ClosureInfo →
    ClosureEmitter.ClosureInfo
  * new ConstructedTypeCollector() → new ClosureEmitter.ConstructedTypeCollector()
  * BodyEmitter callers of IsTaskClrType → ClosureEmitter.IsTaskClrType

ClosureEmitter does NOT take a back-reference to
ReflectionMetadataEmitter — composition only, same pattern as every
other PR-E-* component.

Validation
----------
  * dotnet build GSharp.sln succeeds with 0 errors / 0 warnings.
  * Core.Tests (2120 tests incl. RefactoringBaselineTests,
    DeterministicEmitTests, SlotDictionaryAliasingAssertionTests):
    all pass. Zero hash diffs in RefactoringBaselineTests.
  * Interpreter.Tests (72), LanguageServer.Tests (242),
    Sdk.Tests (24): all pass.
  * Compiler.Tests slices — Closure/Lambda/Async/Issue503/Issue523/
    Issue527 (64 tests), LanguageConformance + Acceptance (83 tests),
    IlVerifier (3 tests): all pass.
  * Full Compiler.Tests run-as-one-process OOMs locally on macOS
    (xUnit testhost heap exhaustion across ~700 compile-and-execute
    tests); GitHub Actions Linux CI runs the suite cleanly — last 5
    main commits all green.

LoC delta
---------
ReflectionMetadataEmitter.cs: 13249 → 12982 (−267)
ClosureEmitter.cs:                 0 → 503 (new)

Public/internal API surface unchanged. No emitted-IL changes.
Sets up: PR-E-10 StateMachineEmitter (next), then PR-E-11
MethodBodyEmitter which lands the #567 / #503 unifying fix on the
structural foundation this PR establishes.
